refactor: use expo-video for picture-in-picture
